Router.reopen with didTransition
