+++ app/initializers/head-tags.js
@@ -1,10 +1,1 @@
-export function initialize() {
-  // ember 1.13 backwards compatibility
-  const application = arguments[1] || arguments[0];
-  application.inject('service:head-tags', 'router', 'router:main');
-}
-
-export default {
-  name: 'head-tags',
-  initialize: initialize
-};
+export { default, initialize } from 'ember-cli-meta-tags/initializers/head-tags';
