+++ app/initializers/router-head-tags.js
@@ -1,0 +1,1 @@
+export { default, initialize } from 'ember-cli-meta-tags/initializers/router-head-tags';
